renamed `exposures.py` to `exposure_tools.py`
