+++ exposure_tools.py
